Tests and pre-commit passes!
