+++ packages/contracts/tests/test_data_schemas.py
@@ -1,4 +1,5 @@
-from datetime import datetime
+from datetime import datetime, timezone
+
 import polars as pl
 import pytest
 from contracts.data_schemas import SubstationFlows
@@ -8,13 +9,11 @@ def test_substation_flows_validation_mw_or_mva():
     # Valid with MW
     df_mw = pl.DataFrame(
         {
-            "substation_name": ["test"],
-            "timestamp": [datetime(2026, 1, 1)],
+            "timestamp": [datetime(2026, 1, 1, tzinfo=timezone.utc)],
             "MW": [10.0],
         }
     ).with_columns(
         [
-            pl.col("substation_name").cast(pl.Categorical),
             pl.col("MW").cast(pl.Float32),
         ]
     )
@@ -25,13 +24,11 @@ def test_substation_flows_validation_mw_or_mva():
     # Valid with MVA
     df_mva = pl.DataFrame(
         {
-            "substation_name": ["test"],
-            "timestamp": [datetime(2026, 1, 1)],
+            "timestamp": [datetime(2026, 1, 1, tzinfo=timezone.utc)],
             "MVA": [10.0],
         }
     ).with_columns(
         [
-            pl.col("substation_name").cast(pl.Categorical),
             pl.col("MVA").cast(pl.Float32),
         ]
     )
@@ -42,13 +39,11 @@ def test_substation_flows_validation_mw_or_mva():
     # Invalid: neither MW nor MVA
     df_none = pl.DataFrame(
         {
-            "substation_name": ["test"],
-            "timestamp": [datetime(2026, 1, 1)],
+            "timestamp": [datetime(2026, 1, 1, tzinfo=timezone.utc)],
             "MVAr": [5.0],
         }
     ).with_columns(
         [
-            pl.col("substation_name").cast(pl.Categorical),
             pl.col("MVAr").cast(pl.Float32),
         ]
     )
@@ -61,14 +56,12 @@ def test_substation_flows_validation_both():
     # Valid with both
     df_both = pl.DataFrame(
         {
-            "substation_name": ["test"],
-            "timestamp": [datetime(2026, 1, 1)],
+            "timestamp": [datetime(2026, 1, 1, tzinfo=timezone.utc)],
             "MW": [10.0],
             "MVA": [12.0],
         }
     ).with_columns(
         [
-            pl.col("substation_name").cast(pl.Categorical),
             pl.col("MW").cast(pl.Float32),
             pl.col("MVA").cast(pl.Float32),
         ]
